set flag to check build status

+++ bot.sh
@@ -4,6 +4,8 @@
 source bot.conf
 source build/envsetup.sh
 
+export BUILD_FINISHED=false
+
 sendMessage() {
 MESSAGE=$1
 
@@ -29,37 +31,41 @@ then
 		echo "build DONEEE"
 		sendMessage "build DONE!"
 
-		#since build iz done lets upload
-		OUTPUT_FILE=$(grep -o -P '(?<=Package\ Complete).*(?=.zip)' build.log)'.zip'
-#		OUTPUT_FILE=$(grep -o -P '(?<=Zip: ).*(?=.zip)' build.log)'.zip'
-		OUTPUT_LOC=$(echo $OUTPUT_FILE | cut -f2 -d":")
-		echo $OUTPUT_LOC
+		BUILD_FINISHED=true
+		if [ $BUILD_FINISHED = true  ] ; then
 
-		#upload to gdrive
-		sendMessage "Uploading to Gdrive..."
-		gdrive upload $OUTPUT_LOC | tee upload.log
+			#since build iz done lets upload
+			OUTPUT_FILE=$(grep -o -P '(?<=Package\ Complete).*(?=.zip)' build.log)'.zip'
+	#		OUTPUT_FILE=$(grep -o -P '(?<=Zip: ).*(?=.zip)' build.log)'.zip'
+			OUTPUT_LOC=$(echo $OUTPUT_FILE | cut -f2 -d":")
+			echo $OUTPUT_LOC
 
-		#get file id
-		sed -e 's/.*Uploaded\(.*\)at.*/\1/' upload.log >> id.txt
-		sendMessage "Upload Finished."
-		sed -e '1d' id.txt >> final.txt
-		FILE_ID=$(cat final.txt)
+			#upload to gdrive
+			sendMessage "Uploading to Gdrive..."
+			gdrive upload $OUTPUT_LOC | tee upload.log
 
-		#set permission to sharing
-		gdrive share $FILE_ID
+			#get file id
+			sed -e 's/.*Uploaded\(.*\)at.*/\1/' upload.log >> id.txt
+			sendMessage "Upload Finished."
+			sed -e '1d' id.txt >> final.txt
+			FILE_ID=$(cat final.txt)
 
-		#finally output the sharing url
-		URL='https://drive.google.com/open?id='$FILE_ID
-		SHARE="$(echo -e "${URL}" | tr -d '[:space:]')"
-		echo $SHARE >> url.txt
+			#set permission to sharing
+			gdrive share $FILE_ID
 
-		#MESSAGE=$(cat url.txt)
-		echo $SHARE
-		sendMessage $SHARE
+			#finally output the sharing url
+			URL='https://drive.google.com/open?id='$FILE_ID
+			SHARE="$(echo -e "${URL}" | tr -d '[:space:]')"
+			echo $SHARE >> url.txt
 
-		BEELD_FINISHED=true
-		#clean up
-		rm upload.log final.txt id.txt url.txt
+			#MESSAGE=$(cat url.txt)
+			echo $SHARE
+			sendMessage $SHARE
+
+			BEELD_FINISHED=true
+			#clean up
+			rm upload.log final.txt id.txt url.txt
+		fi
 	else
 		sendMessage "beeld phel nibba"
 		echo "brunch phel"
@@ -71,8 +77,10 @@ fi
 
 rm lunch.log
 
-	# Some Extra Summary to share
-                MD5=`md5sum ${OUTPUT_LOC} | awk '{ print $1 }'`
+if [ $BUILD_FINISHED = true  ] ; then
+
+# Some Extra Summary to share
+MD5=`md5sum ${OUTPUT_LOC} | awk '{ print $1 }'`
 
 read -r -d '' SUMMARY << EOM
 ROM: $ZIPNAAM
@@ -85,7 +93,7 @@ EOM
                 curl -s "https://api.telegram.org/bot${BOT_API_KEY}/sendmessage" --data "text=$SUMMARY&chat_id=$CHAT_ID" 1> /dev/null
 echo -e;
 
-#fi
+fi
 
 
 exit 1
